progress

+++ projects/batfish/src/main/java/org/batfish/dataplane/ibdp/VirtualRouter.java
@@ -672,7 +672,7 @@ public class VirtualRouter implements Serializable {
         AsPath asPath = advert.getAsPath();
         SortedSet<Long> clusterList = advert.getClusterList();
         SortedSet<Long> communities = ImmutableSortedSet.copyOf(advert.getCommunities());
-        int localPreference = advert.getLocalPreference();
+        long localPreference = advert.getLocalPreference();
         long metric = advert.getMed();
         Prefix network = advert.getNetwork();
         Ip nextHopIp = advert.getNextHopIp();
